Minor changes to files to demonstrate working authentication

+++ src/main/java/com/google/sps/servlets/Entry.java
@@ -5,10 +5,10 @@ public final class Entry {
   private final long id;
   private final String message;
   private final long timestamp;
-  private final float score;
+  private final double score;
   //TODO: consider noting the user too
 
-  public Entry(long id, String message, long timestamp, float score) {
+  public Entry(long id, String message, long timestamp, double score) {
     this.id = id;
     this.message = message;
     this.timestamp = timestamp;
